Removing implicit prefix length now that we support adaptive inlining
